Abstract copy buttons from files
